+++ llm/turbine_llm/ops/__init__.py
@@ -4,5 +4,8 @@
 # See https://llvm.org/LICENSE.txt for license information.
 # SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-from .matmul import *
+from .mmtfp import *
+from .mmt_block_scaled_offset_q4 import *
+from .mmt_block_scaled_q8 import *
+from .mmt_super_block_scaled_offset_q4 import *
 from .custom_inference_ops import CustomInferenceOps
